requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
DateTime==4.3
idna==3.3
kiwisolver==1.3.2
mariadb==1.0.7
matplotlib==3.4.3
mysql-connector-python==8.0.27
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
protobuf==3.19.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
zope.interface==5.4.0
